fix: improve style
