+++ src/ui/pages/Portfolio/PortfolioSubmenu/SubmenuActionButtons/WithdrawMarginDetails/WithdrawMarginDetails.styled.tsx
@@ -1,5 +1,23 @@
 import styled from '@emotion/styled';
 
+export const WithdrawMarginDetailsWrapperBox = styled('div')`
+  box-sizing: border-box;
+
+  display: flex;
+  flex-direction: column;
+  gap: 24px;
+  width: 100%;
+`;
+
+export const WithdrawMarginDetailsBox = styled('div')`
+  box-sizing: border-box;
+
+  display: flex;
+  flex-direction: column;
+  gap: 8px;
+  width: 100%;
+`;
+
 export const TransactionDetailBox = styled('div')`
   box-sizing: border-box;
